move the Custom instruction into the same list
